Support multiple windows and sort tabs
If you open a new window from an existing console window (e.g. with Ctrl-Shift-N), then this would previously end up in the same session file.
This results in all the tabs in a single window after restore. This commit puts every window in a seperate file.

Also: Tabs are now sorted properly. If you change the order of tabs (drag and drop), they keep their session id. This results in a wrong restore.
To fix this, we use "/Windows/x sessionList", which gives the right order.